pyproject.toml
[project]
name = "pamir-mlpermafrost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "lukegre", email = "lukegre@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "cryogrid-pytools>=0.3.18",
    "dask>=2025.7.0",
    "geopandas>=1.1.1",
    "gpytorch>=1.14",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.1.7",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "rioxarray",
    "s3fs>=2025.7.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "xarray>=2025.7.1",
    "zarr>=3.1.0",
]

[project.scripts]
pamir-mlpermafrost = "pamir_mlpermafrost:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
